+++ src/tui/framework/store.ts
@@ -1,4 +1,3 @@
-import { includesEqual } from '@whimbrel/array'
 import { deletePath, PropertyPath, readPath, writePath } from '@whimbrel/walk'
 import equal from 'fast-deep-equal'